// my-app/src/models/Constants.tsx
export enum SkillStatus {
  FOUND = "Found",
  MISSING = "Missing",
  PARTIALLY_FOUND = "Partially Found",
}

export enum SkillImportance {
  HIGH = "High",
  MEDIUM = "Medium",
  LOW = "Low",
}

export enum RecommendationType {
  CRITICAL = "Critical",
  IMPROVEMENT = "Improvement",
  OPTIMIZATION = "Optimization",
  ENHANCEMENT = "Enhancement",
}

export enum RecommendationImpact {
  HIGH = "High",
  MEDIUM = "Medium",
  LOW = "Low",
}